fix: enforce compile-time format string validation and resolve cross-platform defects
Introduced the ION_FORMAT_PRINTF compiler attribute macro to restore strict
GCC/Clang format string validation for custom variadic functions (e.g., _isprintf)
to resolve CodeQL alert #186. Enabling this validation unmasked legacy type-safety
issues and formatting defects across the codebase, which have now been resolved:

- Global/Cross-Platform: Migrated volatile size_t, time_t, PsmAddress, and SdrObject
  specifiers to portable UVAST_FIELDSPEC and ADDR_FIELDSPEC macros with explicit
  casting, guaranteeing type safety across 32-bit and 64-bit deployments.
- BPv7: Fixed 32/64-bit desyncs for IPN service numbers in ewchar logging,
  corrected lookUpEndpoint mismatches, and resolved admin CLI formatting.
- BPSec & Crypto: Fixed format-extra-args (trailing NULLs), missing parentheses,
  and uvast casting in BPSEC, BCB, BIB, and CSI debug macros.
- Network Management (NM): Swept agent, mgr, utils, and primitives to fix
  AMP_DEBUG trailing NULLs, heapmax types, pthread_self() casting, and a
  critical logic bug in iif_is_registered.
- CFDP & DTPC: Corrected integer specifier mismatches (%d vs %ld) in dtpcsend
  and prevented an SdrObject segfault in the CFDP module.
- BSS, TC & IPND: Resolved size_t and time_t mismatches while preserving
  legacy compiler support for IPND.
- ICI Core: Fixed a macro typo (TCP_LOW_CYCLE instead of TCPCL_LOW_CYCLE) in
  platform.c that prevented the inclusion of <netinet/tcp.h>, resolving a
  TCP_SYNCNT undeclared identifier error when building with --enable-tcpcl-low-cycle.
- Testing & Style: Resolved format errors across the test suites and applied
  clang-format to all strict validation patches.

Verified cleanly against default, BPSec, ASAN, and comprehensive kitchen-sink
configurations.

+++ ici/sdr/sdrxn.c
@@ -2573,9 +2573,10 @@ static int	writeToLog(const char *file, int line, Sdr sdrv, char *from,
 		{
 			char buf[256];
 			isprintf(buf, sizeof(buf), "Log max size exceeded. \
-SDR: %s  logSize: %lu logLength: %lu length: %lu depth: %d", sdr->name,
-					sdr->logSize, sdr->logLength,
-					length, sdr->xnDepth);
+SDR: %s  logSize: %lu logLength: %lu length: %lu depth: %d",
+					sdr->name, (unsigned long) sdr->logSize,
+					(unsigned long) sdr->logLength,
+					(unsigned long) length, sdr->xnDepth);
 			_putErrmsg(file, line, buf, NULL);
 			return -1;
 		}
